refactor(authRepo): change structure
